+++ app/assets/javascripts/app/pages/contacts.js
@@ -22,6 +22,8 @@ app.pages.Contacts = Backbone.View.extend({
 
     this.aspectCreateView = new app.views.AspectCreate({ el: $("#newAspectContainer") });
     this.aspectCreateView.render();
+
+    this.setupAspectSorting();
   },
 
   toggleChatPrivilege: function() {
@@ -73,6 +75,21 @@ app.pages.Contacts = Backbone.View.extend({
 
   searchContactList: function(e) {
     this.stream.search($(e.target).val());
+  },
+
+  setupAspectSorting: function() {
+    $("#aspect_nav .nav").sortable({
+      items: "li.aspect[data-aspect-id]",
+      update: function() {
+        $("#aspect_nav .ui-sortable").removeClass("synced");
+        var data = JSON.stringify({ ordered_aspect_ids: $(this).sortable("toArray", { attribute: "data-aspect-id" }) });
+        $.ajax(Routes.order_aspects_path(),
+          { type: "put", dataType: "text", contentType: "application/json", data: data })
+          .done(function() { $("#aspect_nav .ui-sortable").addClass("synced"); });
+      },
+      revert: true,
+      helper: "clone"
+    });
   }
 });
 // @license-end