+++ spec/models/aspect_spec.rb
@@ -39,6 +39,13 @@ describe Aspect, :type => :model do
     it 'has a contacts_visible? method' do
       expect(alice.aspects.first.contacts_visible?).to be true
     end
+
+    it "sets an order_id" do
+      aspect_2 = alice.aspects.create(name: "People")
+      expect(aspect_2.order_id).to eq(2)
+      aspect_3 = alice.aspects.create(name: "Cat People")
+      expect(aspect_3.order_id).to eq(3)
+    end
   end
 
   describe 'validation' do